requirements.txt
alembic==1.6.5
blinker==1.4
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
cryptography==3.4.7
dnspython==2.1.0
dominate==2.6.0
email-validator==1.1.2
Flask==2.0.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.0.1
Flask-Moment==1.0.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.0
gTTS==2.2.2
http-ece==1.1.0
idna==2.9
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.4
MarkupSafe==2.0.1
numpy==1.20.3
opencv-python==4.5.2.54
Pillow==8.2.0
psycopg2==2.8.6
py-vapid==1.8.2
pycparser==2.20
PyJWT==2.1.0
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
pywebpush==1.13.0
reportlab==3.5.67
requests==2.25.1
six==1.16.0
SQLAlchemy==1.4.18
urllib3==1.26.5
visitor==0.1.3
Werkzeug==2.0.1
WTForms==2.3.3
